[Calendar] Fix the guild-filter reader and park the rest behind one reply
CMSG_CALENDAR_GUILD_FILTER (0x04E3) read three uint32s where the client writes
three BYTES. Its writer sub_668A7F calls sub_40F018 three times from object
offsets +16, +17 and +18, so the whole body is three bytes and the inherited
reader asked for twelve. The widths are right on their own terms too: two
character levels and a guild rank index, none of which can exceed 255.

The opcode is not registered, so nothing dispatched it and the over-read was
unreachable. It is fixed now rather than later because the rebuild has to precede
registration, not follow it.

Also proven and needing no change: CMSG_CALENDAR_REMOVE_EVENT (0x0C61) and
CMSG_CALENDAR_COPY_EVENT (0x1A97) both read uint64 eventId, uint64 inviteId,
uint32 -- which is exactly what sub_665987 writes. They legitimately share one
writer; the body shape is identical and each has its own thunk.

The other twelve calendar CMSGs stay dormant, and the reason is recorded in
Opcodes.cpp: every one of them reaches CalendarMgr::SendCalendarCommandResult,
and SMSG_CALENDAR_COMMAND_RESULT (0x142A) is still the legacy body -- uint32(0),
uint8(0), a cstring param or uint8(0), uint32(err) -- neither converted nor
admitted by the send gate. Registering any of them would answer with that body,
which is the CMSG_CONTACT_LIST mistake repeated.

Worth recording because it is not visible from the handler: three of the twelve
emit no SMSG in their own body and look free. GUILD_FILTER, REMOVE_EVENT and
COPY_EVENT all reach the command result one level down, inside
CalendarMgr::RemoveEvent, CalendarMgr::CopyEvent and Guild::MassInviteToEvent. A
handler-only scan clears them; following the call does not.

CMSG_CALENDAR_ADD_EVENT additionally needs a full reader rewrite. sub_66D4E2 is
bit-packed -- four uint32 and a byte, then a bit-packed invite count and
description length, per-invite 8-bit GUID masks, a title length, a flush, then the
GUID bytes and both strings LAST. The current reader reads title-first with a
pre-MoP ReadAsPacked() GUID and bears no resemblance to it.

Also folds in the one MINOR from the GLM-5.2 review of b26db31f0: the
SMSG_CONTACT_LIST pre-allocation hint still assumed 25 bytes per entry. It is now
34, the online-friend width, with the real widths recorded. The review put the
minimum at 29; it is 21, as the fixture asserts. Hint only -- WorldPacket grows on
demand either way. That review returned APPROVE with no blocking or important
findings, having independently decompiled the client readers and writers.

Build clean, 115/115 ctest green.

+++ src/game/WorldHandlers/CalendarHandler.cpp
@@ -212,9 +212,16 @@ void WorldSession::HandleCalendarGuildFilter(WorldPacket& recv_data)
 {
     DEBUG_LOG("WORLD: Received opcode CMSG_CALENDAR_GUILD_FILTER [%s]", _player->GetGuidStr().c_str());
 
-    uint32 minLevel;
-    uint32 maxLevel;
-    uint32 minRank;
+    // Three BYTES, not three uint32s. The client's writer sub_668A7F emits
+    // sub_40F018 three times from object offsets +16, +17 and +18, so the whole
+    // body is three bytes. The inherited reader asked for twelve and would have
+    // over-read the buffer on every request.
+    //
+    // The widths are also right on their own terms: all three are a character
+    // level or a guild rank index, and none can exceed 255.
+    uint8 minLevel;
+    uint8 maxLevel;
+    uint8 minRank;
 
     recv_data >> minLevel >> maxLevel >> minRank;
